# pkg/apiserver/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "Resource for managing API Resources",
    "title": "Supported Resources API",
    "version": "v0.0.0"
  },
  "paths": {
    "/apis/resources/v1/actions": {
      "get": {
        "consumes": [
          "*/*"
        ],
        "produces": [
          "application/json",
          "application/yaml",
          "application/vnd.kubernetes.protobuf"
        ],
        "summary": "list or watch objects of kind Action",
        "operationId": "func1",
        "parameters": [
          {
            "type": "string",
            "name": "labelSelector",
            "in": "query"
          },
          {
            "type": "string",
            "name": "fieldSelector",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "watch",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "allowWatchBookmarks",
            "in": "query"
          },
          {
            "type": "string",
            "name": "resourceVersion",
            "in": "query"
          },
          {
            "type": "string",
            "name": "resourceVersionMatch",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "timeoutSeconds",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "name": "continue",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "sendInitialEvents",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "progressNotify",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/apis.ActionList"
            }
          }
        }
      },
      "post": {
        "consumes": [
          "*/*"
        ],
        "produces": [
          "application/json",
          "application/yaml",
          "application/vnd.kubernetes.protobuf"
        ],
        "summary": "create an Action",
        "operationId": "func4",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/apis.Action"
            }
          },
          {
            "type": "string",
            "name": "dryRun",
            "in": "query"
          },
          {
            "type": "string",
            "name": "fieldManager",
            "in": "query"
          },
          {
            "type": "string",
            "name": "fieldValidation",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/apis.Action"
            }
          },
          "201": {
            "description": "Created",
            "schema": {
              "$ref": "#/definitions/apis.Action"
            }
          }
        }
      },
      "delete": {
        "consumes": [
          "*/*"
        ],
        "produces": [
          "application/json",
          "application/yaml",
          "application/vnd.kubernetes.protobuf"
        ],
        "summary": "delete collection of Action",
        "operationId": "func7",
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/meta.DeleteOptions"
            }
          },
          {
            "type": "integer",
            "name": "gracePeriodSeconds",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "orphanDependents",
            "in": "query"
          },
          {
            "type": "string",
            "name": "propagationPolicy",
            "in": "query"
          },
          {
            "type": "string",
            "name": "dryRun",
            "in": "query"
          },
          {
            "type": "string",
            "name": "labelSelector",
            "in": "query"
          },
          {
            "type": "string",
            "name": "fieldSelector",
            "in": "query"
          },
          {
            "type": "string",
            "name": "resourceVersion",
            "in": "query"
          },
          {
            "type": "string",
            "name": "resourceVersionMatch",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "timeoutSeconds",
            "in": "query"
          },
          {
            "type": "integer",
            "name": "limit",
            "in": "query"
          },
          {
            "type": "string",
            "name": "continue",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "sendInitialEvents",
            "in": "query"
          },
          {
            "type": "boolean",
            "name": "progressNotify",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/definitions/v1.Status"
            }
          }
        }
      }
    },
...
  },
  "definitions": {
    "apis.Action": {
      "required": [
        "Spec",
        "Status"
      ],
      "properties": {
        "Spec": {
          "$ref": "#/definitions/apis.ActionSpec"
        },
        "Status": {
          "$ref": "#/definitions/apis.ActionStatus"
        },
        "apiVersion": {
          "type": "string"
        },
        "creationTimestamp": {
          "$ref": "#/definitions/meta.Time"
        },
        "deletionGracePeriodSeconds": {
          "type": "integer",
          "format": "int64"
        },
        "deletionTimestamp": {
          "$ref": "#/definitions/meta.Time"
        },
        "generateName": {
          "type": "string"
        },
        "generation": {
          "type": "integer",
          "format": "int64"
        },
        "kind": {
          "type": "string"
        },
        "labels": {
          "type": "object",
          "additionalProperties": {
            "type": "string"
          }
        },
        "name": {
          "type": "string"
        },
        "namespace": {
          "type": "string"
        },
        "resourceVersion": {
          "type": "string"
        },
        "uid": {
          "type": "string"
        }
      }
    },
    "apis.ActionList": {
      "required": [
        "Items"
      ],
      "properties": {
        "Items": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.Action"
          }
        },
        "apiVersion": {
          "type": "string"
        },
        "continue": {
          "type": "string"
        },
        "kind": {
          "type": "string"
        },
        "remainingItemCount": {
          "type": "integer",
          "format": "int64"
        },
        "resourceVersion": {
          "type": "string"
        }
      }
    },
    "apis.ActionSpec": {
      "required": [
        "Name",
        "Parents",
        "Desc",
        "Type",
        "Conditions",
        "Runtimes",
        "EnableFineGrainedControl"
      ],
      "properties": {
        "Conditions": {
          "$ref": "#/definitions/apis.Conditions"
        },
        "Desc": {
          "$ref": "#/definitions/apis.Description"
        },
        "EnableFineGrainedControl": {
          "type": "boolean"
        },
        "Name": {
          "type": "string"
        },
        "Parents": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Runtimes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.Runtime"
          }
        },
        "Type": {
          "type": "string"
        }
      }
    },
    "apis.ActionStatus": {
      "required": [
        "ActionID",
        "Belongs",
        "Phase",
        "Resources",
        "Devices",
        "Data",
        "Scenes",
        "RuntimeStatus",
        "Results",
        "StartAt",
        "FinishAt",
        "LastTime"
      ],
      "properties": {
        "ActionID": {
          "type": "string"
        },
        "Belongs": {
          "$ref": "#/definitions/apis.IDRef"
        },
        "Data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.DataStatus"
          }
        },
        "Devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.DeviceStatus"
          }
        },
        "FinishAt": {
          "$ref": "#/definitions/apis.Time"
        },
        "LastTime": {
          "$ref": "#/definitions/apis.Time"
        },
        "Phase": {
          "type": "string"
        },
        "Resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.ResourceStatus"
          }
        },
        "Results": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.Result"
          }
        },
        "RuntimeStatus": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.RuntimeStatus"
          }
        },
        "Scenes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.SceneStatus"
          }
        },
        "StartAt": {
          "$ref": "#/definitions/apis.Time"
        }
      }
    },
    "apis.ConditionFormula": {
      "required": [
        "LeftValue",
        "RightValue",
        "Signal",
        "Join"
      ],
      "properties": {
        "Join": {
          "type": "string"
        },
        "LeftValue": {
          "$ref": "#/definitions/apis.ConditionValue"
        },
        "RightValue": {
          "$ref": "#/definitions/apis.ConditionValue"
        },
        "Signal": {
          "type": "string"
        }
      }
    },
...
    "apis.Conditions": {
      "required": [
        "Formulas"
      ],
      "properties": {
        "Formulas": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.ConditionFormula"
          }
        }
      }
    },
...
    "apis.DataSpec": {},
    "apis.DataStatus": {},
    "apis.Description": {
      "required": [
        "Label",
        "Docs"
      ],
      "properties": {
        "Docs": {
          "type": "string"
        },
        "Label": {
          "type": "array",
          "items": {
            "type": "string"
          }
        }
      }
    },
    "apis.DeviceSpec": {},
    "apis.DeviceStatus": {},
...
    "apis.IDRef": {
      "required": [
        "WorkflowID",
        "TaskID",
        "GroupID",
        "ActionID"
      ],
      "properties": {
        "ActionID": {
          "type": "string"
        },
        "GroupID": {
          "type": "string"
        },
        "TaskID": {
          "type": "string"
        },
        "WorkflowID": {
          "type": "string"
        }
      }
    },
...
    "apis.ResourceSpec": {},
    "apis.ResourceStatus": {},
    "apis.Result": {
      "required": [
        "ResultID",
        "Belongs",
        "Type"
      ],
      "properties": {
        "Belongs": {
          "$ref": "#/definitions/apis.IDRef"
        },
        "ResultID": {
          "type": "string"
        },
        "Type": {
          "type": "string"
        }
      }
    },
    "apis.Runtime": {
      "required": [
        "Resources",
        "Devices",
        "Data",
        "Scenes",
        "Type",
        "Name",
        "Image",
        "Command",
        "Args",
        "EnvVar",
        "Inputs",
        "Outputs"
      ],
      "properties": {
        "Args": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Command": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "Data": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.DataSpec"
          }
        },
        "Devices": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.DeviceSpec"
          }
        },
        "EnvVar": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.EnvVar"
          }
        },
        "Image": {
          "type": "string"
        },
        "Inputs": {
          "$ref": "#/definitions/apis.Input"
        },
        "Name": {
          "type": "string"
        },
        "Outputs": {
          "$ref": "#/definitions/apis.Output"
        },
        "Resources": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.ResourceSpec"
          }
        },
        "Scenes": {
          "type": "array",
          "items": {
            "$ref": "#/definitions/apis.SceneSpec"
          }
        },
        "Type": {
          "type": "string"
        }
      }
    },
    "apis.RuntimeStatus": {
      "required": [
        "NodeName",
        "ProcessId",
        "Phase",
        "StartAt",
        "FinishAt",
        "LastTime"
      ],
      "properties": {
        "FinishAt": {
          "$ref": "#/definitions/apis.Time"
        },
        "LastTime": {
          "$ref": "#/definitions/apis.Time"
        },
        "NodeName": {
          "type": "string"
        },
        "Phase": {
          "type": "string"
        },
        "ProcessId": {
          "type": "string"
        },
        "StartAt": {
          "$ref": "#/definitions/apis.Time"
        }
      }
    },
    "apis.SceneSpec": {},
    "apis.SceneStatus": {},
...
    "apis.Time": {
      "required": [
        "Time"
      ],
      "properties": {
        "Time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
...
    "meta.DeleteOptions": {
      "properties": {
        "apiVersion": {
          "type": "string"
        },
        "dryRun": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "gracePeriodSeconds": {
          "type": "integer",
          "format": "int64"
        },
        "kind": {
          "type": "string"
        },
        "orphanDependents": {
          "type": "boolean"
        },
        "preconditions": {
          "$ref": "#/definitions/meta.Preconditions"
        },
        "propagationPolicy": {
          "type": "string"
        }
      }
    },
    "meta.Preconditions": {
      "properties": {
        "resourceVersion": {
          "type": "string"
        },
        "uid": {
          "type": "string"
        }
      }
    },
    "meta.Time": {
      "required": [
        "Time"
      ],
      "properties": {
        "Time": {
          "type": "string",
          "format": "date-time"
        }
      }
    },
    "v1.ListMeta": {
      "description": "ListMeta describes metadata that synthetic resources must have, including lists and various status objects. A resource may have only one of {ObjectMeta, ListMeta}.",
      "properties": {
        "continue": {
          "description": "continue may be set if the user set a limit on the number of items returned, and indicates that the server has more data available. The value is opaque and may be used to issue another request to the endpoint that served this list to retrieve the next set of available objects. Continuing a consistent list may not be possible if the server configuration has changed or more than a few minutes have passed. The resourceVersion field returned when using this continue value will be identical to the value in the first response, unless you have received this token from an error message.",
          "type": "string"
        },
        "remainingItemCount": {
          "description": "remainingItemCount is the number of subsequent items in the list which are not included in this list response. If the list request contained label or field selectors, then the number of remaining items is unknown and the field will be left unset and omitted during serialization. If the list is complete (either because it is not chunking or because this is the last chunk), then there are no more remaining items and this field will be left unset and omitted during serialization. Servers older than v1.15 do not set this field. The intended use of the remainingItemCount is *estimating* the size of a collection. Clients should not rely on the remainingItemCount to be set or to be exact.",
          "type": "integer",
          "format": "int64"
        },
        "resourceVersion": {
          "description": "String that identifies the server's internal version of this object that can be used by clients to determine when objects have changed. Value must be treated as opaque by clients and passed unmodified back to the server. Populated by the system. Read-only. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#concurrency-control-and-consistency",
          "type": "string"
        },
        "selfLink": {
          "description": "Deprecated: selfLink is a legacy read-only field that is no longer populated by the system.",
          "type": "string"
        }
      }
    },
    "v1.Status": {
      "description": "Status is a return value for calls that don't return other objects.",
      "properties": {
        "apiVersion": {
          "description": "APIVersion defines the versioned schema of this representation of an object. Servers should convert recognized schemas to the latest internal value, and may reject unrecognized values. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#resources",
          "type": "string"
        },
        "code": {
          "description": "Suggested HTTP return code for this status, 0 if not set.",
          "type": "integer",
          "format": "int32"
        },
        "details": {
          "description": "Extended data associated with the reason.  Each reason may define its own extended details. This field is optional and the data returned is not guaranteed to conform to any schema except that defined by the reason type.",
          "$ref": "#/definitions/v1.StatusDetails"
        },
        "kind": {
          "description": "Kind is a string value representing the REST resource this object represents. Servers may infer this from the endpoint the client submits requests to. Cannot be updated. In CamelCase. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "message": {
          "description": "A human-readable description of the status of this operation.",
          "type": "string"
        },
        "metadata": {
          "description": "Standard list metadata. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "$ref": "#/definitions/v1.ListMeta"
        },
        "reason": {
          "description": "A machine-readable description of why this operation is in the \"Failure\" status. If this value is empty there is no information available. A Reason clarifies an HTTP status code but does not override it.",
          "type": "string"
        },
        "status": {
          "description": "Status of the operation. One of: \"Success\" or \"Failure\". More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#spec-and-status",
          "type": "string"
        }
      }
    },
    "v1.StatusCause": {
      "description": "StatusCause provides more information about an api.Status failure, including cases when multiple errors are encountered.",
      "properties": {
        "field": {
          "description": "The field of the resource that has caused this error, as named by its JSON serialization. May include dot and postfix notation for nested attributes. Arrays are zero-indexed.  Fields may appear more than once in an array of causes due to fields having multiple errors. Optional.\n\nExamples:\n  \"name\" - the field \"name\" on the current resource\n  \"items[0].name\" - the field \"name\" on the first array entry in \"items\"",
          "type": "string"
        },
        "message": {
          "description": "A human-readable description of the cause of the error.  This field may be presented as-is to a reader.",
          "type": "string"
        },
        "reason": {
          "description": "A machine-readable description of the cause of the error. If this value is empty there is no information available.",
          "type": "string"
        }
      }
    },
    "v1.StatusDetails": {
      "description": "StatusDetails is a set of additional properties that MAY be set by the server to provide additional information about a response. The Reason field of a Status object defines what attributes will be set. Clients must ignore fields that do not match the defined type of each attribute, and should assume that any attribute may be empty, invalid, or under defined.",
      "properties": {
        "causes": {
          "description": "The Causes array includes more details associated with the StatusReason failure. Not all StatusReasons may provide detailed causes.",
          "type": "array",
          "items": {
            "$ref": "#/definitions/v1.StatusCause"
          }
        },
        "group": {
          "description": "The group attribute of the resource associated with the status StatusReason.",
          "type": "string"
        },
        "kind": {
          "description": "The kind attribute of the resource associated with the status StatusReason. On some operations may differ from the requested resource Kind. More info: https://git.k8s.io/community/contributors/devel/sig-architecture/api-conventions.md#types-kinds",
          "type": "string"
        },
        "name": {
          "description": "The name attribute of the resource associated with the status StatusReason (when there is a single name which can be described).",
          "type": "string"
        },
        "retryAfterSeconds": {
          "description": "If specified, the time in seconds before the operation should be retried. Some errors may indicate the client must take an alternate action - for those errors this field may indicate how long to wait before taking the alternate action.",
          "type": "integer",
          "format": "int32"
        },
        "uid": {
          "description": "UID of the resource. (when there is a single resource which can be described). More info: https://kubernetes.io/docs/concepts/overview/working-with-objects/names#uids",
          "type": "string"
        }
      }
    }
  }
}